requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
parse_pip_search==0.0.1
pillow==12.1.0
Pygments==2.19.2
PyJWT==2.11.0
python-decouple==3.8
razorpay==2.0.0
requests==2.32.5
rich==14.2.0
soupsieve==2.8.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
